edit : namespace edit & ChronoTime del  - C++17 버전 namespace 를 이전 버전 namespace로 변경 및 ChronoTime을 제외

+++ WindowEngine_SOURCE/CAnimatorComponent.cpp
@@ -101,7 +101,7 @@ namespace Framework
 				vecImgs[0]->GetHDC(), 0, 0, SRCCOPY);
 
 		}
-		CreateAnimation(name, spriteSheet, Vector2::Zero, Vector2(imgWidth, imgHeigth), offset, fileCount, duration);
+		CreateAnimation(name, spriteSheet, Vector2::Zero, Vector2((float)imgWidth, (float)imgHeigth), offset, fileCount, duration);
 	}
 
 	CAnimation* CAnimatorComponent::FindAnimation(const std::wstring& name)
